chore: refactor commands
